requirements necrew.txt
streamlit>=1.33
pytesseract
pillow
pandas
gspread
oauth2client
streamlit-authenticator
